+++ mozrepl/mozrepl.py
@@ -9,13 +9,12 @@ import uuid
 import urlparse
 import urllib
 import tempfile
+import base64
+import json
 #from ufp.terminal.debug import print_ as debug
 
 from .exception import Exception as MozException
 from .type import Object, Function, Array
-
-DEFAULT_HOST = '127.0.0.1'
-DEFAULT_PORT = 4242
 
 class Mozrepl(object):
 	"""
@@ -32,16 +31,25 @@ class Mozrepl(object):
 	:type host: unicode
 	"""
 	_RE_PROMPT = re.compile(r'^repl\d*>', re.MULTILINE)
+	DEFAULT_HOST = '127.0.0.1'
+	DEFAULT_PORT = 4242
 	
 	def __init__(self, port=DEFAULT_PORT, host=DEFAULT_HOST):
 		"""
 		mozrepl Firefox Add-on과 연결합니다.
 		"""
 		self.connect(port, host)
+		
+		self._baseVarname = '__pymozrepl_{0}'.format(uuid.uuid4().hex)
+		
+		buffer = """var {baseVar} = {{ 'ref': {{}}, 'context': {{}} }}; null;""".format(baseVar=self._baseVarname)
+		self._rawExecute(buffer)
 	
 	def connect(self, port=None, host=None):
 		"""
 		mozrepl Firefox Add-on과 연결합니다.
+		
+		연결 대상은 최초 연결된 대상과 같아야 합니다.
 		
 		:param port: mozrepl Firefox Add-on의 포트. 생략시, 기존 값을 사용합니다.
 		:param port: mozrepl Firefox Add-on의 호스트. 생략시, 기존 값을 사용합니다.
@@ -56,17 +64,6 @@ class Mozrepl(object):
 		
 		match = self._telnet.expect([self._RE_PROMPT], 10)[1]
 		self.prompt = match.group(0)
-		
-		self._baseVarname = '__pymozrepl_{0}'.format(uuid.uuid4().hex)
-		
-		self._rawExecute("""\
-				var {baseVar} = {{
-					'ref': Object(),
-					'lastCmdValue': null
-				}}
-				""".format(baseVar=self._baseVarname)
-			)
-		pass
 	
 	def __repr__(self):
 		return 'Mozrepl(port={port}, host={host})'.format(port=repr(self.port), host=repr(self.host))
@@ -97,109 +94,84 @@ class Mozrepl(object):
 		:return: Firefox MozREPL Add-on에서 응답이 없는 경우 None을 반환.
 		:rtype: unicode
 		"""
-		self._telnet.write('{command};\n'.format(command=command).encode('utf8')) #전송
-		respon = self._telnet.read_until(self.prompt).decode('utf8') #수신
+		#전송
+		buffer = """try {{ {command}; }} catch (e) {{ (function() {{ let robj = {{ 'exception': null }}; Object.getOwnPropertyNames(e).forEach(function (key) {{ robj.exception[key] = e[key]; }}, e); let buffer = JSON.stringify(robj); buffer = window.btoa(unescape(encodeURIComponent(buffer))); return robj; }}()) }};""".format(command=command)
+		self._telnet.write(buffer.encode('UTF-8'))
+		
+		respon = self._telnet.read_until(self.prompt).decode('UTF-8') #수신
 		
 		#아무 응답도 없을 경우 None을 반환
 		if re.search('^ %(prompt)s$' % vars(self), respon):
 			return None
 		
 		#응답이 존재하는 경우 응답받은 문자열에서 불필요한 문자열을 제거.
-		respon = re.sub('^ ', '', respon) #응답된 문자열의 앞 공백 제거
-		respon = re.sub(r'\n%(prompt)s$' % vars(self), '', respon, re.UNICODE) #입력 프롬프트 제거
+		respon = re.sub('^ "', '', respon) #응답된 문자열의 앞 공백 제거
+		respon = re.sub(r'"\n%(prompt)s$' % vars(self), '', respon, re.UNICODE) #입력 프롬프트 제거
+		
+		respon = base64.decodestring(respon).decode('UTF-8')
+		respon = json.loads(respon) #받은 내용을 파싱.
 		
 		#오류일 경우 예외를 던짐.
-		match = re.match(r'!{3} \[\S+? (?P<typeNmae>\S+?)\]\n\nDetails:\n(?P<details>.*?)\n', respon, re.DOTALL) \
-			or re.match(r'!{3} (?P<typeNmae>\S+?): (?P<summary>.+?)\n\nDetails:\n(?P<details>.*?)\n$', respon, re.DOTALL) \
-			or re.match(r'!{3} \[Exception\.{3} (?P<summary>"Component returned failure code: 0x8000ffff \((?P<typeNmae>\S+?)\).*?)\]\n\nDetails:\n(?P<details>.*?)\n$', respon, re.DOTALL) \
-			or re.match(r'!{3} \[Exception\.{3} (?P<summary>.*?)\]\n\nDetails:\n(?P<details>.*?)\n$', respon, re.DOTALL) \
-			or re.match(r'!{3} (?P<summary>.*)\nDetails:\n(?P<details>.*?)\n$', respon, re.DOTALL) \
-			or re.match(r'!{3} (?P<summary>.*)$', respon, re.DOTALL)
-		if match:
-			groupdict = match.groupdict()
-			typeName = groupdict.get('typeNmae', '')
-			summary = groupdict.get('summary', '')
-			details = groupdict.get('details', '')
-			raise MozException(typeName, summary, details)
+		if 'exception' in respon:
+			raise MozException(respon['exception'])
 		
 		return respon
 	
 	def __del__(self):
-		self._rawExecute("""
-			delete {baseVar}.ref
-			delete {baseVar}
-			""".format(baseVar=self._baseVarname)
-			)
-		pass
+		buffer = """delete {baseVar}; null;""".format(baseVar=self._baseVarname)
+		self._rawExecute(buffer)
 	
 	def execute(self, command):
 		"""
 		명령을 실행합니다.
 		
-		.. attention:: 오브젝트의 메소드를 사용할때, 'repl.execute("repl.home")()'과 같이 함수의 메소드를 바로 반환 받은 뒤 사용 할 수 없습니다. 이와 같은 방식으로 사용하려면, 'repl.execute("repl.home").call("repl")'와 같이 호출 오브젝트를 명시적으로 넘겨주거나, 'repl.execute("repl").home()'과 같이 pymozrepl의 자동 바인딩 기능을 사용하십시오.
+		.. attention:: 오브젝트의 메소드를 사용할때, 'repl.execute("repl.home")()'과 같이 함수의 메소드를 바로 반환 받은 뒤 사용 할 수 없습니다. 이와 같은 방식으로 사용하려면, 'repl.execute("repl.home.call")(Raw("repl"))'와 같이 호출 오브젝트를 명시적으로 넘겨주거나, 'repl.execute("repl").home()'과 같이 pymozrepl의 자동 바인딩 기능을 사용하십시오.
+		.. attention:: 하나의 pymozrepl 객체는 독립된 하나의 문맥을 가집니다. 이 문맥은 Firefox mozrepl에서 제공하는 현재 문맥과는 다르다는 점을 주의하십시오.
+		.. todo:: 명령 응답 중에 prompt가 섞여서 일부 내용은 완전히 받지 못 할 수 있다. 또, -inf나 long 따위의 값은 처리를 하지 못하는 문제가 있다. 이런 문제를 해결하기 위해, 응답을 json으로 만들고 파일 입출력을 통해 반환받도록 한다.
 		
 		:param command: 명령.
 		:type command: unicode
 		:raise mozrepl.Exception: mozrepl Firefox Add-on에서 오류를 던질 경우.
-		:returns: int : mozrepl Firefox Add-on에서 반환받은 값이 정수인 경우.
-		:returns: float : mozrepl Firefox Add-on에서 반환받은 값이 실수인 경우.
-		:returns: unicode : mozrepl Firefox Add-on에서 반환받은 값이 문자열인 경우.
 		:returns: :py:class:`~mozrepl.type.Object` : mozrepl Firefox Add-on에서 반환받은 값이 object인 경우.
 		:returns: :py:class:`~mozrepl.type.Array` : mozrepl Firefox Add-on에서 반환받은 값이 array인 경우.
 		:returns: :py:class:`~mozrepl.type.Function` : mozrepl Firefox Add-on에서 반환받은 값이 function인 경우.
-		:returns: unicode : mozrepl Firefox Add-on에서 반환받은 값을 분석 할 수 없는 경우 또는 type이 repr로  설정된 경우에 응답 받은 값을 그대로 반환합니다.
-		:returns: bool : mozrepl Firefox Add-on에서 반환받은 값이 진리형인 경우.
+		:returns: 그 외에 string, number, bool 등의 기본값은 python에서 대응되는 적절한 기본 타입(int, bool, unicode 등)으로 변환하여 돌려줍니다.
 		"""
 		#명령을 실행
-		with tempfile.NamedTemporaryFile('wb', prefix='.tmp_pymozrepl_', suffix='.js') as tempFile:
-			tempFile.write(command.encode('UTF-8'))
-			tempFile.flush()
-			scriptUrl = urlparse.urljoin('file:', urllib.pathname2url(tempFile.name.encode('UTF-8')))
-			respon = self._rawExecute('{baseVar}.lastCmdValue = repl.loader.loadSubScript("{scriptUrl}", this, "UTF-8")'.format(scriptUrl=scriptUrl, baseVar=self._baseVarname)) #명령을 mozrepl 서버에 전송
+		with tempfile.NamedTemporaryFile('wb', prefix='.tmp_pymozrepl_', suffix='.js') as scriptFile:
+			#명령 기록
+			scriptFile.write(command.encode('UTF-8'))
+			scriptFile.flush()
+			
+			#기록된 명령 파일의 주소를 생성
+			buffer = scriptFile.name.encode('UTF-8')
+			buffer = urllib.pathname2url(buffer)
+			scriptUrl = urlparse.urljoin('file:', buffer)
+			
+			#명령을 mozrepl 서버에 전송
+			buffer = """(function(){{ let robj = {{}}; let lastCmdValue = repl.loader.loadSubScript("{scriptUrl}", {baseVar}.context, "UTF-8"); robj.type = typeof lastCmdValue; if ( robj.type == 'object' ) {{ if ( Array.isArray(lastCmdValue) ) {{ robj.type = 'array'; }}; {baseVar}.ref['{refUuid}'] = lastCmdValue; robj.refUuid = '{refUuid}'; }} else {{ robj.value = lastCmdValue; }}; let buffer = JSON.stringify(robj); buffer = window.btoa(unescape(encodeURIComponent(buffer))); return buffer; }}());""".format(
+				scriptUrl = scriptUrl, 
+				baseVar = self._baseVarname,
+				refUuid = uuid.uuid4()
+				)
+			respon = self._rawExecute(buffer)
 		
 		#응답받은 결과가 없으면 그대로 반환
 		if respon is None:
 			return None
 		
-		#boolean
-		if respon in ['false', 'true']:
-			return respon == 'true'
-		
-		#Number - int
-		if re.match(r'\d+$', respon):
-			return int(respon)
-		
-		#Number - float
-		if re.match(r'\d+\.\d+$', respon):
-			return float(respon)
-		
-		#string
-		match = re.match(r'"(.*)"$', respon)
-		if match:
-			return match.group(1)
-		
 		#function
-		if respon == 'function() {...}':
-			uuid_ = unicode(uuid.uuid4())
-			match = re.search('(\S+?)(\.[^.]+|\[.+?\])\s*$', command)
-			self._rawExecute('{baseVar}.ref["{uuid}"] = {baseVar}.lastCmdValue'.format(baseVar=self._baseVarname, uuid=uuid_))
-			return Function(self, uuid_)
+		if respon['type'] == 'function':
+			return Function(self, respon['refUuid'])
 		
-		#타입 분석.
-		buffer = Object.makeNotinited(self)
-		type = self.execute("""{buffer} = {baseVar}.lastCmdValue; typeof {baseVar}.lastCmdValue""".format(baseVar=self._baseVarname, buffer=buffer))
+		#array
+		if respon['type'] == 'array':
+			return Array(self, respon['refUuid'])
 		
-		#array, object
-		if type == 'object':
-			uuid_ = unicode(uuid.uuid4())
-			self._rawExecute('{baseVar}.ref["{uuid}"] = {buffer}'.format(baseVar=self._baseVarname, buffer=buffer, uuid=uuid_))
-			
-			#array
-			if self.execute('Array.isArray({buffer})'.format(buffer=buffer)):
-				return Array(self, uuid_)
-			
-			#object
-			return Object(self, uuid_)
+		#object
+		if respon['type'] == 'object':
+			return Object(self, respon['refUuid'])
 		
-		return respon
+		#기본 타입
+		return respon['value']